Rework to pull Grimoire ns listing from site
This patch removes the fixed list of documented namespaces in
clojure.core, and replaces it with a dynamically generated list of all
the namespaces in the latest version of Clojure on Grimoire.

This still doesn't fix the underlying issue of not being able to match
arbitrary symbols to documentation URLs, but `clojure.core` is pretty
much the only thing on Grimoire right now so we'll go with this as a
start.